schemas and HTTPexceptions added

+++ users/crud.py
@@ -23,17 +23,24 @@ class Crud():
         
                     
         async def get_user_by_email(self,email: str)->schemas.User:
-            return await database.fetch_one(users.select().where(users.c.email == email))
+            user = await database.fetch_one(users.select().where(users.c.email == email))
+            if user is None:
+                 raise HTTPException(status_code=404, detail="User not found")
+            return user
+            
         
 
         async def get_user_by_id(self,id: int)->schemas.User:
             user = await database.fetch_one(users.select().where(users.c.id == id))
             if user is None:
-                 raise HTTPException(status_code=404, detail="User not found")
+                 raise HTTPException(status_code=404, detail=f"User with id {id} not found")
             return user
 
         async def update_user(self,id:int,user:schemas.UserUpdate)->schemas.User:
             now = datetime.utcnow()
+            user = await database.fetch_one(users.select().where(users.c.id == id))
+            if user is None:
+                 raise HTTPException(status_code=404, detail=f"User with id {id} not found")
             query = users.update().where(users.c.id == id).values(
             first_name = user.first_name,
             phone_number=user.phone_number,
@@ -44,6 +51,9 @@ class Crud():
             return schemas.User(**user.dict(),id=id)
 
         async def delete_user(self,id:int)->str:
+            user = await database.fetch_one(users.select().where(users.c.id == id))
+            if user is None:
+                 raise HTTPException(status_code=404, detail=f"User with id {id} not found")
             query = users.delete().where(users.c.id == id)
             await database.execute(query)
             return {f"User with id {id} is successfully deleted"}
@@ -51,4 +61,4 @@ class Crud():
 
         async def get_all_users(self,skip: int = 0, limit: int = 100)->List[schemas.User]:
             results = await database.fetch_all(users.select().offset(skip).limit(limit))
-            return [dict(result) for result in results]
+            return [schemas.User(**result) for result in results]
